image converter & decoder resize to renderer size, etc
resizing in ImageConverter is supported, we compare qt resizing with hwa
or ipp (color convertion and )resizing.
before this, resizing in ImageConverter will cause a crash when resizing
the renderer, because of the data of image changes in video thread when
it is painting in main thread. But locking is not enough, because the
data passed to renderer is ref which (address) may changed in video
thread's ImageConverter, so a copy of that data is required.