Disconnect sourcePlayers signal stopped from vfilter/afilter when stopping AVTranscoder.
The missing disconnect leads to a second call to finish of vfilter/afilter
which then leads to a segmentation fault in sync mode. In async mode
the segmentation fault does not happen because of a race condition
although a second call is executed. Always stop AVTranscoder first and
then the associated AVPlayer.