Split AudioFrames according to the audio encoders frame_size
Some encoders (e.g. OPUS) define a frame_size for each audio frame
and only process that much data in one encoding step. This means that
the number of incoming frames does not have be equal to the number of
outgoing frames. This patch splits AudioFrames according to the audio
encoders frame_size and buffers the leftovers of a frame until the next
frame arrives and prepends them to it. This results in the decoupling of
incoming and outgoing frame sizes.