fix seek updates video while paused
This fixes #1149 that was introduced with #1119. The previous fix had moved the clearing of packet queues to inside the SeekTask::run. But if the thread has been paused, the SeekTask would not run. This puts back the packetQueue clearing in AVDemuxThread::seek. And keeps the packet clearing also in SeekTask::run